Finally working Form with JSTL:
 - Inserting new Products in Database
 - Retrieving them back;
 - Listing Products in JSP with JSTL

+++ Spring/src/br/spring/mvc/model/daos/ProductDAO.java
@@ -1,6 +1,5 @@
 package br.spring.mvc.model.daos;
 
-import java.util.ArrayList;
 import java.util.List;
 
 import javax.persistence.EntityManager;
@@ -8,6 +7,8 @@ import javax.persistence.PersistenceContext;
 import javax.persistence.TypedQuery;
 import javax.transaction.Transactional;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Repository;
 
 import br.spring.mvc.model.entities.Product;
@@ -19,55 +20,98 @@ public class ProductDAO {
 	@PersistenceContext
 	private EntityManager aEntityManager;
 	
+	private Logger logger;
+	
+	public ProductDAO() {
+		// Logging messages
+		this.logger = LoggerFactory.getLogger("spring.mvc.model.daos.ProductDAO");
+		this.logger.info("ProductDAO instancialized");
+	}
+	
+	/**
+	 * Insert's a new Product in the database
+	 * 
+	 * @param pProduto
+	 */
 	public void save(Product pProduto) {
 		this.aEntityManager.persist(pProduto);
 	}
 	
+	/**
+	 * Select all Products existing in database
+	 * 
+	 * @return		A Product list
+	 */
 	public List<Product> selectAll() {
-		List<Product> resultSet = new ArrayList<Product>();
-		
+		// A general Select query using JPQL
 		String sql = "Select p From Product p";
 		
+		// A TypedQuery is created to do a general Select at database
 		TypedQuery<Product> typedQuery = 
 			this.aEntityManager.createQuery(sql, Product.class);
 		
-		resultSet = typedQuery.getResultList();
+		// The ResultSet is ready
+		List<Product> resultSet = typedQuery.getResultList();
+		
 		
 		return resultSet;
 	}
 	
+	/**
+	 * Delete all Products existing in database
+	 */
 	public void deleteAll() {
-		List<Product> resultSet = new ArrayList<Product>();
+		// A List with all Product is selected at database
+		List<Product> resultSet = selectAll();
 		
-		String sql = "Select p From Product p";
-		
-		TypedQuery<Product> typedQuery = 
-			this.aEntityManager.createQuery(sql, Product.class);
-		
-		resultSet = typedQuery.getResultList();
-		
+		// For each Product existing in database, Delete them one by one
 		for (Product product : resultSet) {
 			this.aEntityManager.remove(product);
 		}
 	}
 	
+	/**
+	 * Select a specific Products, if it exists in database
+	 * 
+	 * @return		A Product list
+	 */
 	public Product select(Product pProduct) {
-		Product result = 
-			this.aEntityManager.find(Product.class, pProduct.getPrimaryKey());
+		Product result = null;
 		
-		if ( result != null ) {
-			this.aEntityManager.remove(result);
+		// Validate the Product parameter, before trying to use it, avoiding a NullPointerException
+		if (
+				pProduct 					!= null		&& 
+				pProduct.hasAnyAttributeValid()
+		) {
+			// Executes an Primary Key selection in database
+			result = 
+				this.aEntityManager.find(
+					Product.class, 
+					pProduct.getPrimaryKey()
+				);
+			
+			/* TODO Why this peace of code does an Delete?!
+			 *
+			if ( result != null ) {
+				this.aEntityManager.remove(result);
+			}
+			*/
 		}
 		
 		return result;
 	}
 	
+	/**
+	 * Delete a specific Products, if it exists in database
+	 * 
+	 * @param pProduct
+	 */
 	public void delete(Product pProduct) {
+		// Selects a Product, if it exists in database
 		Product result = 
-			this.aEntityManager.find(
-				Product.class, 
-				pProduct.getPrimaryKey());
+			select(pProduct);
 		
+		// If the Product were found at database, Delete it.
 		if ( result != null ) {
 			this.aEntityManager.remove(result);
 		}
